Extract HttpEventStore exceptions to a separate file for greater readability.
